refactor(common): change postSlice name to postsSlice and updated imports
